Revert "Refactor code into modules"
