+++ test/test.js
@@ -3,22 +3,19 @@
   const fs = require('fs');
   const vm = require('vm');
 
-  const pluginCode = fs.readFileSync('dist/plugin.js', 'utf8');
-  const suggestCode = fs.readFileSync('dist/suggest.js', 'utf8');
-  const holidaysCode = fs.readFileSync('dist/holidays.js', 'utf8');
-
-  const funcSrc = holidaysCode.match(/function phraseToMoment\([^]*?\n\}/);
+  const code = fs.readFileSync('main.js', 'utf8');
+  const funcSrc = code.match(/function phraseToMoment\([^]*?\n\}/);
   if (!funcSrc) throw new Error('phraseToMoment not found');
-  const classSrc = suggestCode.match(/class DDSuggest[^]*?\n\}/);
+  const classSrc = code.match(/class DDSuggest[^]*?\n\}/);
   if (!classSrc) throw new Error('DDSuggest class not found');
-  const pluginSrc = pluginCode.match(/class DynamicDates[^]*?\n\}/);
+  const pluginSrc = code.match(/class DynamicDates[^]*?\n\}/);
   if (!pluginSrc) throw new Error('DynamicDates class not found');
-  const settingsSrc = pluginCode.match(/exports\.DEFAULT_SETTINGS\s*=\s*{[^]*?};/);
+  const settingsSrc = code.match(/const DEFAULT_SETTINGS =[^]*?};/);
   if (!settingsSrc) throw new Error('DEFAULT_SETTINGS not found');
-  const helpersSrc = holidaysCode.match(/function nthWeekdayOfMonth[^]*?function needsYearAlias[^]*?function isHolidayQualifier[^]*?function formatTypedPhrase[^]*?\nexports\.PHRASES/);
+  const helpersSrc = code.match(/function nthWeekdayOfMonth[^]*?function needsYearAlias[^]*?function isHolidayQualifier[^]*?function formatTypedPhrase[^]*?\nconst PHRASES/);
   if (!helpersSrc) throw new Error('helper functions not found');
   const helpersCode = helpersSrc[0]
-    .replace(/exports\.DEFAULT_SETTINGS[^]*?};/, '')
+    .replace(/const DEFAULT_SETTINGS[^]*?};/, '')
     .replace(/\nconst PHRASES[^]*/, '');
 
   /* ------------------------------------------------------------------ */
@@ -110,14 +107,12 @@
   };
   const context = { moment, WEEKDAYS, MONTHS, BASE_WORDS, EditorSuggest, KeyboardEvent, Plugin, PluginSettingTab, Setting, obsidian_1 };
   vm.createContext(context);
-  vm.runInContext('var exports = this;', context);
   vm.runInContext('this.MONTH_ABBR = this.MONTHS.map(m => m.slice(0,3));', context);
   vm.runInContext('this.expandMonthName = function(name){ const idx = this.MONTH_ABBR.indexOf(name.slice(0,3).toLowerCase()); return idx >= 0 ? this.MONTHS[idx] : name; };', context);
   vm.runInContext(helpersCode, context);
   vm.runInContext('this.HOLIDAY_PHRASES = HOLIDAY_PHRASES;', context);
   vm.runInContext('this.PHRASES = this.BASE_WORDS.flatMap(w => this.WEEKDAYS.includes(w) ? [w, "last " + w, "next " + w] : [w]).concat(this.HOLIDAY_PHRASES.flatMap(h => [h, "last " + h, "next " + h]));', context);
   vm.runInContext(funcSrc[0], context);
-  vm.runInContext('var exports = this; var holidays_1 = exports; var suggest_1 = { default: this.DDSuggest }; function require(m){ if(m==="obsidian") return obsidian_1; if(m==="./holidays") return holidays_1; if(m==="./suggest") return suggest_1; return {}; }', context);
   vm.runInContext(settingsSrc[0], context);
   vm.runInContext('this.DDSuggest=' + classSrc[0], context);
   vm.runInContext('this.DynamicDates=' + pluginSrc[0], context);
